requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
python-multipart==0.0.6
jinja2==3.1.2
python-dotenv==1.0.0
pydantic==2.4.2
asyncio-mqtt==0.16.1
apscheduler==3.10.4
